+++ examples/wordpress-theme/bud.config.js
@@ -1,5 +1,12 @@
 module.exports = app =>
   app
+    .use([
+      require('@roots/bud-postcss'),
+      require('@roots/bud-wordpress-dependencies'),
+      require('@roots/bud-wordpress-externals'),
+      require('@roots/bud-wordpress-manifests'),
+      require('@roots/bud-entrypoints'),
+    ])
     .when(app.isDevelopment, app =>
       app.use([
         require('@roots/bud-babel'),
@@ -12,13 +19,6 @@ module.exports = app =>
         .hash()
         .runtime('single'),
     )
-    .use([
-      require('@roots/bud-postcss'),
-      require('@roots/bud-wordpress-dependencies'),
-      require('@roots/bud-wordpress-externals'),
-      require('@roots/bud-wordpress-manifests'),
-      require('@roots/bud-entrypoints'),
-    ])
     .proxy()
     .entry({
       app: ['app.js', 'app.css'],