+++ packages/@roots/bud-build/src/Rule/index.ts
@@ -69,8 +69,10 @@ class Rule implements Contract {
   }
 
   @bind
-  public setExclude(exclude: (app: Framework) => RegExp) {
-    this.exclude = exclude
+  public setExclude(
+    exclude: RegExp | ((app: Framework) => RegExp),
+  ) {
+    this.exclude = isFunction(exclude) ? exclude : () => exclude
   }
 
   @bind