+++ packages/@roots/bud-tailwindcss/src/index.ts
@@ -1,27 +1,15 @@
-import './interface'
-import {Framework, Module} from '@roots/bud-framework'
+import {Tailwind} from './interface'
 import {tailwind} from './api'
 
-export const name: Module['name'] = '@roots/bud-tailwindcss'
-
-export const api: Module['api'] = {tailwind}
-
-export const boot: Module['boot'] = (app: Framework) => {
-  /**
-   * Exit early if requirements not met
-   */
-  if (
-    !app.discovery.has('devDependencies.postcss') ||
-    !app.discovery.has('devDependencies.tailwindcss')
-  ) {
-    return
-  }
-
-  const implementation = app.discovery.has(
-    'devDependencies.@tailwindcss/jit',
-  )
-    ? '@tailwindcss/jit'
-    : 'tailwindcss'
-
-  app.tailwind(null, implementation)
+const tailwindcss: Tailwind.Extension = {
+  name: '@roots/bud-tailwindcss',
+  api: {
+    tailwind,
+  },
+  boot: app => {
+    app.tailwind()
+  },
 }
+
+export const {name, api, boot} = tailwindcss
+export default tailwindcss