+++ packages/@roots/sage/src/index.ts
@@ -1,2 +1,6 @@
-export {name, boot} from './sage/index'
+import {sage} from './sage'
+
+export default sage
+export const {name, boot} = sage
+
 export type {Sage} from './sage/interface'