+++ tests/util/index.ts
@@ -1,36 +1,6 @@
-import {
-  Bud,
-  Framework,
-  config as defaultConfig,
-  services,
-} from '@roots/bud'
-import {Signale} from 'signale'
-
-const logger = new Signale({
-  types: {
-    log: {
-      badge: '📝',
-      label: 'log',
-      color: 'white',
-    },
-    error: {
-      badge: '🚨',
-      label: 'error',
-      color: 'red',
-    },
-    success: {
-      badge: '✅',
-      label: 'success',
-      color: 'green',
-    },
-  },
-})
-logger.config({
-  displayFilename: true,
-  displayTimestamp: false,
-  displayDate: false,
-})
-const {log, error, success} = logger
+import {Bud, config as defaultConfig, services} from '@roots/bud'
+import {Framework} from '@roots/bud-framework'
+import {logger, log, success, error} from './logger'
 
 /**
  * On the annoying failure to link asm message...
@@ -39,15 +9,12 @@ const {log, error, success} = logger
  * https://github.com/vadimdemedes/yoga-layout-prebuilt/issues/8
  * https://github.com/diegomura/react-pdf/issues/603
  */
-
-const config = {...defaultConfig}
-
 const setupBud = (
   modeOverride?: 'development' | 'production',
   configOverride?: any,
   servicesOverride?: any,
 ) => {
-  let bud: Framework = new Bud(configOverride ?? config)
+  const bud: Framework = new Bud(configOverride ?? defaultConfig)
   bud.mode = modeOverride ?? 'production'
 
   bud.bootstrap(servicesOverride ?? services)
@@ -65,15 +32,18 @@ const teardownBud = (bud: Framework) => {
 }
 
 const checkState = <T = any>(state: T) => {
-  if (state) error('bud state persisted between tests', state)
+  state && error('bud state persisted between tests', state)
 }
+
+export {helper as integration, Assets} from './integration'
 
 export {
   Bud,
   Framework,
   setupBud,
   teardownBud,
-  config,
+  defaultConfig as config,
+  logger,
   log,
   error,
   success,